+++ client/src/index.css
@@ -6,6 +6,7 @@ body {
   min-height: 100vh;
   width: 100vw;
   overflow-x: hidden;
+  position: relative;
 }
 
 /* Mobile responsiveness */
